transaction basic saving

+++ src/main/java/org/akshay/dto/SplitAmount.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Data
 public class SplitAmount {
     User user;
-    float amount;
+    double amount;
 }
